Move TeeZip state out of reducing function
Sadly, the IR test for `Zip(Count(), xf_double)` is broken now (so
commented out).  There is another broken inference test marked with
at-test_broken.

However, it's not clear if this is a particular problem in TeeZip.  It
seems that this is problematic already in simple nested transducers.
For example, following example cannot be inferred (even with `init=1`;
checked in 1.1.1 and 1.3.0-DEV.508)

    foldl(*, Scan(+) |> Scan(+), 1:3)

Interestingly, manual type assertion fixes the issue

    foldl(*, Scan(+) |> Map(x -> x::Int) |> Scan(+), 1:3)

So, there is a hope for solving this issue in a more automatic way
than forcing users to invoke type inference manually via `eduction`.