Associative string-splitting transducer (#344)
Fix the word-splitting transducer in `examples/words.jl` to avoid
re-ordering words.  This assumes the inner reducing function to be
associative.  In turn, it makes the whole composite reducing function
"associative" (in the sense that `combine` works).

Making it work requires `collect` to use a proper monoid (i.e.,
`start(rf, Init)` works).