Tail-call function-barrier for arrays (#403)
This PR implements the "tail-call function-barrier" pattern for arrays
with linear style indexing.  This gives us a better performance for
type-changing reduction where the iteration at which the type changes
is unknown.  A good example is filtered sum with no initial value
(#404).

Interestingly, this version is a bit worse for sum of filtered floats
on "short" arrays.  Compare the result of
`["filter_sum", "1000", "RandomFloats", "noinit", "xf"]`.  In my
laptop, baseline is ~1.7 μs while this branch is ~2.1 μs.  However,
for 10x longer input, this branch is faster (> 2x).  The benchmarks on
CI also show similar results.