Add ReducePartitionBy (#458)
This patch adds a parallelizable transducer `ReducePartitionBy(f, rf, [init])`
that can be used for fusing intra- and inter-partition reductions. It also
is used for a performance specialization of `array |> PartitionBy(f)` that
does not require allocations of vectors any more.

For example, `ReducePartitionBy` can be used for implementing group reduce
operations that runs in parallel and uses constant space (Note: hash-based
group reduce requires the space at least as large as/proportional to the
number of groups).