More uniform reduce chunk sizes
It's usually desirable to have chunk sizes as similar as possible. If we simply divide the array by the number of threads, sometimes the last chunk can turn out to have a very different size from the others. By changing the way we compute the indices, dividing by the number of arrays after we multiply by the chunk index, we ensure chunk sizes differ at most by 1.